module4-langchain/requirements.txt
langchain
langchain-openai
langchain-google-genai
langchain_experimental
openai
anthropic
cohere
streamlit
python-dotenv
beautifulsoup4
faiss-cpu
pypdf
unstructured
networkx
openpyxl
rapidocr-onnxruntime
